feat: add zipcode

+++ src/Bemol.Customer.Api/Integrations/External/ViaCep/Interfaces/IViaCepService.cs
@@ -4,6 +4,6 @@ namespace Bemol.Customer.Api.Integrations.External.ViaCep.Interfaces
 {
     public interface IViaCepService
     {
-        Task<ZipCodeResponse> GetZipCode(int zipCode);
+        Task<ZipCodeResponse> GetZipCode(long zipCode);
     }
 }
